[48619] Fixed APITokens::CreateContract spec
